requirements.txt
pdfplumber==0.11.0
PyPDF2
Flask
Werkzeug==2.3.8
pywebview
pywebview
